switch back to old log reporter when "verbose" or "grouped" are used
