+++ packages/reporters/src/index.ts
@@ -2,9 +2,6 @@ export { AdoReporter } from "./AdoReporter.js";
 export { JsonReporter } from "./JsonReporter.js";
 export { LogReporter } from "./LogReporter.js";
 export { ChromeTraceEventsReporter } from "./ChromeTraceEventsReporter.js";
-
-export { initializeReporters } from "./initialize.js";
-export { createReporter } from "./createReporter.js";
+export { ProgressReporter } from "./ProgressReporter.js";
 
 export type { TargetStatusEntry, TargetMessageEntry } from "./types/TargetLogEntry.js";
-export type { ReporterInitOptions } from "./initialize.js";
